move:#196 select-box > 공통 ui로 이동
