+++ src/components/ui/select-box.tsx
@@ -1,12 +1,10 @@
 import { DEFAULT } from '@/constants/user-meta-data-constants';
 import type { SelectBoxType } from '@/types/select-box';
 
-//@TODO:
-// 사람인 api 연결 시 현재 select box의 선택값이 name으로 되어있지만 value로 수정할 것
 type Props = {
   options: SelectBoxType[];
   selected: string;
-  onSelect: (value: SelectBoxType['value']) => void; //@TODO: 추후 type 'value'로 수정
+  onSelect: (value: SelectBoxType['value']) => void;
 };
 
 const SelectBox = ({ options, selected = DEFAULT, onSelect }: Props) => {
